backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-chess==1.999
stockfish==3.28.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.7

# LLM Support
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.13

# RAG & ML Support
chromadb==0.4.15
sentence-transformers
scikit-learn==1.3.2

# Data & Web
pandas==2.1.2
plotly==5.17.0
